added .httr-oauth
Directly added .httr-oauth hidden file that was created on CWK computer to access EI Google Sheet as this seems to enable the full R code to write and read to the Google Sheet but since the .gitignore file was preventing this from getting uploaded (or even other things prevent the upload) this authorization was not embedded in the built Docker image. So this is a test to see if building the Docker image with this file will now enable the Docker containerized version of EIoF code to acces EI Google Sheet without asking for a web browser to then authenticate.